pyproject.toml
[tool.poetry]
name = "demo-mobail-autotest-selene"
version = "0.1.0"
description = ""
authors = ["Andrey.Ch"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.4.2"
allure-pytest = "^2.13.2"
appium-python-client = "2.6.1"
pillow = "^10.0.1"
selene = "2.0.0-beta.8"
requests = "2.28.1"
pytest-rerunfailures = "^12.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
